Logica de usuario premium deleteProduct

+++ src/controllers/product.controller.js
@@ -26,6 +26,7 @@ export const getProductById = async (req, res) => {
 export const addProduct = async (req, res) => {
   try {
     const product = req.body;
+    product.owner = req._id;
     const result = await ProductService.createProduct(product);
 
     return res.json({ result });
@@ -51,9 +52,22 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   try {
     const { pid } = req.params;
-    const result = await ProductService.deleteProduct(pid);
+    const { role, _id } = req;
+    console.log({role});
 
-    return res.json({ result });
+    if (role === 'premium') {
+      const product = await ProductService.getProductById(pid);
+      if (!product) return res.status(404).json({ msg: `El producto con id ${pid} no existe` });
+
+      if (product.owner.toString() === _id) {
+        const result = await ProductService.deleteProduct(pid);
+        if (result) return res.json({ success: true, msg: 'Producto eliminado', result });
+        return res.status(400).json({ success: false, msg: 'No se pudo eliminar el producto con id: ' + pid });
+      }
+    }
+    const result = await ProductService.deleteProduct(pid);
+    if (result) return res.json({ success: true, msg: 'Producto eliminado', result });
+    return res.status(400).json({ success: false, msg: 'No se pudo eliminar el producto con id: ' + pid });
   } catch (error) {
     console.log('deleteProduct error: ', error);
     return res.status(500).json({ msg: 'Internal server error' });
